Fix for GCP-790
git-svn-id: https://efficio.atlassian.net/svn/GCP/trunk/IBPMiddleware@2354 4ebb2e00-0138-4608-bc6e-edf3948d4863

+++ src/main/java/org/generationcp/middleware/dao/gdms/MarkerDAO.java
@@ -614,7 +614,7 @@ public class MarkerDAO extends GenericDAO<Marker, Integer>{
     }
     
     @SuppressWarnings("rawtypes")
-    public Set<Integer> getMarkerIDsByMapIDAndLinkageBetweenStartPosition(int mapID, String linkageGroup, int startPos, int endPos, int start, int numOfRows) throws MiddlewareQueryException{
+    public Set<Integer> getMarkerIDsByMapIDAndLinkageBetweenStartPosition(int mapID, String linkageGroup, double startPos, double endPos, int start, int numOfRows) throws MiddlewareQueryException{
         try {
             
             SQLQuery query;
@@ -637,7 +637,7 @@ public class MarkerDAO extends GenericDAO<Marker, Integer>{
     }
     
     @SuppressWarnings("rawtypes")
-    public long countMarkerIDsByMapIDAndLinkageBetweenStartPosition(int mapID, String linkageGroup, int startPos, int endPos) throws MiddlewareQueryException{
+    public long countMarkerIDsByMapIDAndLinkageBetweenStartPosition(int mapID, String linkageGroup, double startPos, double endPos) throws MiddlewareQueryException{
         try {
             
             SQLQuery query;
